+++ tests/encryption.test.js
@@ -3,7 +3,7 @@ const {
     EncryptionSettings,
     HistoryVisibility,
     VerificationState,
-} = require("../pkg/matrix_sdk_crypto_js");
+} = require("../pkg/matrix_sdk_crypto_wasm");
 
 describe("EncryptionAlgorithm", () => {
     test("has the correct variant values", () => {
