+++ tsconfig.json
@@ -3,7 +3,7 @@
         "strict": true
     },
     "typedocOptions": {
-        "entryPoints": ["pkg/matrix_sdk_crypto_js.d.ts"],
+        "entryPoints": ["pkg/matrix_sdk_crypto_wasm.d.ts"],
         "out": "docs",
         "readme": "README.md"
     }
